resources/functions/discord/requirements.txt
pynacl==1.5.0
requests==2.27.1
aws-xray-sdk==2.14.0